// src/state/User/user.action-types.ts
//types
export enum ActionType {
  //SIGN IN
  EMAIL_SIGN_IN_START = 'EMAIL_SIGN_IN_START',
  SIGN_IN_SUCCESS = 'SIGN_IN_SUCCESS',
  SIGN_IN_ERROR = 'SIGN_IN_ERROR',
  //SIGN OUT
  EMAIL_SIGN_OUT_START = 'EMAIL_SIGN_OUT_START',
  SIGN_OUT_SUCCESS = 'SIGN_OUT_SUCCESS',
  //USER SESSION
  CHECK_USER_SESSION = 'CHECK_USER_SESSION',
  //SIGN UP
  EMAIL_SIGN_UP_START = 'EMAIL_SIGN_UP_START',
  SIGN_UP_SUCCESS = 'SIGN_UP_SUCCESS',
  SIGN_UP_ERROR = 'SIGN_UP_ERROR',
  //PASSWORD RECOVERY
  PASSWORD_RECOVERY_START = 'PASSWORD_RECOVERY_START',
  PASSWORD_RECOVERY_SUCCESS = 'PASSWORD_RECOVERY_SUCCESS',
  PASSWORD_RECOVERY_ERROR = 'PASSWORD_RECOVERY_ERROR',
  //RESET FORMS
  RESET_AUTH_FORMS = 'RESET_AUTH_FORMS',
}